--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,9 @@ exclude = [
 [lib]
 name = "amplify"
 path = "src/lib.rs"
+# Allow --save-baseline to work
+# https://github.com/bheisler/criterion.rs/issues/275
+bench = false
 
 [dependencies]
 libc = { version = "0.2", optional = true }
@@ -48,7 +51,7 @@ stringly_conversions = { version = "0.1.1", optional = true, features = [
 
 # avoid building criterion in 1.41.1 CI
 [target.'cfg(bench)'.dev-dependencies]
-criterion = "0.2.11"
+criterion = "0.4"
 softposit = "0.3.9"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]