benchmark comparison with 'critcmp':

```
group                     patched                                master
-----                     -------                                ------
i1024_add                 1.00     46.5±2.35ns        ? ?/sec    1.20     55.6±3.58ns        ? ?/sec
i1024_cmp_eq              1.00     15.8±0.91ns        ? ?/sec    1.40     22.3±1.16ns        ? ?/sec
i1024_cmp_mixed           1.00     10.1±0.38ns        ? ?/sec    1.73     17.5±0.66ns        ? ?/sec
i1024_cmp_pos             1.00     15.8±0.88ns        ? ?/sec    1.41     22.3±0.91ns        ? ?/sec
i1024_is_negative         1.00      3.5±0.12ns        ? ?/sec    4.68     16.2±0.56ns        ? ?/sec
i1024_is_positive         1.00     15.8±0.84ns        ? ?/sec    1.03     16.3±0.84ns        ? ?/sec
i1024_zero_is_negative    1.00     16.4±0.61ns        ? ?/sec    1.02     16.8±0.61ns        ? ?/sec
i1024_zero_is_positive    1.00     16.6±1.02ns        ? ?/sec    1.02     16.9±0.83ns        ? ?/sec
posit32_add               1.00     32.2±3.37ns        ? ?/sec    1.00     32.2±1.80ns        ? ?/sec
posit32_div               1.00     49.0±1.75ns        ? ?/sec    1.01     49.6±2.37ns        ? ?/sec
posit32_mul               1.00     26.2±1.42ns        ? ?/sec    1.00     26.3±1.11ns        ? ?/sec
posit32_sub               1.02     34.3±3.57ns        ? ?/sec    1.00     33.5±1.83ns        ? ?/sec
softposit32_add           1.04      9.9±0.38ns        ? ?/sec    1.00      9.5±0.39ns        ? ?/sec
softposit32_div           1.00     22.6±0.85ns        ? ?/sec    1.01     22.8±0.82ns        ? ?/sec
softposit32_mul           1.00      8.6±0.35ns        ? ?/sec    1.27     10.9±0.50ns        ? ?/sec
softposit32_sub           1.05     12.3±0.56ns        ? ?/sec    1.00     11.6±0.51ns        ? ?/sec
u1024_add                 1.00     32.1±1.93ns        ? ?/sec    1.00     32.0±1.04ns        ? ?/sec
u1024_cmp_eq              1.00      8.2±0.48ns        ? ?/sec    1.24     10.2±0.38ns        ? ?/sec
u1024_div                 1.00   296.9±13.45ns        ? ?/sec    1.00   297.5±16.29ns        ? ?/sec
u1024_mul                 1.00   416.2±23.71ns        ? ?/sec    1.28   533.3±16.83ns        ? ?/sec
u1024_sub                 1.00     42.1±1.41ns        ? ?/sec    1.00     42.0±1.64ns        ? ?/sec
```

All these tests were run on my laptop, in a not very controlled environment, so your mileage may vary. But the big differences in the i1024 tests, and in `u1024_cmp_eq`, are quite robust. There's a big and robust improvement in `u1024_mul` too, although I don't fully understand where that's coming from.